Add ChartViewSlice file and its reducer to store
