+++ src/components/OngoingIncidents/OngoingIncidents.js
@@ -10,11 +10,12 @@ const OngoingIncidents = () => {
   const dispatch = useDispatch();
 
   useEffect(() => {
-    dispatch(getOngoingIncidents())
+    dispatch(getOngoingIncidents())//req adds results to state
   }, [dispatch])
 
   const ongoingIncidents = useSelector(state => state.ongoingIncidents.ongoingIncidents).map(incident => {
     return (
+
       <div className='incident' key={incident.id}>
         <h3>{incident.attributes.name}</h3>
         <p>{incident.attributes.type} * DECLARED: {incident.attributes.start_date}</p>
