+++ src/components/ResolvedIncidents/ResolvedIncidents.css
@@ -1,4 +1,4 @@
 .resolved-container {
-  max-height: 30vh;
+  max-height: 22vh;
   overflow-y: scroll;
 }
