+++ src/components/app/App.js
@@ -21,10 +21,13 @@ const App = () => {
           <Route path='/profiles' component={Profiles} />
           <Route path='/passport' component={Passport} />
           {/* below is test-only for specific incident view */}
-          <Route path='/test' component={ChartView} />
-          <Route path='/incident/:id' render={({ match }) => {
+          <Route path='/incident-details/:id' render={({ match }) => {
             const { id } = match.params;
             return (<IncidentInfo id={id} />)
+          }} />
+          <Route path='/incident/:id' render={({ match }) => {
+            const { id } = match.params;
+            return (<ChartView id={id} />)
           }} />
         </Switch>
         <Footer />
